requirements.txt
Flask==2.2.2
Flask_RESTful==0.3.9
pandas==1.4.4
python_google_places==1.4.2
xmltodict==0.13.0
